New structure for different kinds of command/event handlers
There is some experimentation here, to write command/event handlers in a
way that allows them to be more portable/less naieve. Traditionally
event handlers become quite bloated, and this framework attempts to
harness that and add a bit of rationality.

The 'Collection Request' (aka `request`) and 'Campaign' aggregates are
now complete and commuicate with appropriate aggregate
backends. Migrations etc are all included.